Move `trackers` under `tapioca::runtime`
