+++ lib/tapioca/compilers/gem/listeners/sorbet_required_ancestors.rb
@@ -12,7 +12,7 @@ module Tapioca
 
           sig { override.params(event: ScopeNodeAdded).void }
           def on_scope(event)
-            ancestors = Trackers::RequiredAncestor.required_ancestors_by(event.constant)
+            ancestors = Runtime::Trackers::RequiredAncestor.required_ancestors_by(event.constant)
             ancestors.each do |ancestor|
               next unless ancestor # TODO: We should have a way to warn from here
               event.node << RBI::RequiresAncestor.new(ancestor.to_s)
